Truncation of integer division to integers in serial_z85.py and z85.py to make code compatible with python 3.

+++ serial_z85.py
@@ -21,7 +21,7 @@ def receive_and_decode_Z85(ser,size):
     if (size % 4):
         raise ValueError("size should be multiple of 4.")
 
-    size_encoded = size*5/4
+    size_encoded = int(size*5/4)
 
     z85bytes = ser.read(size_encoded)
     # ser.reset_input_buffer()
